Fix crash in debug builds when rendering gothic text with recent fonts
It may happen that text with surrogates is subject to kerning, in which case
we go through the OpenType path. After glyph substitution we reconstruct the
log clusters, but that code currently doesn't work with surrogates.
Currently glyph substitution does not happen with surrogates, so disabling
the logcluster recreation if we didn't do glyph substitution in the first place
works around it.

Signed-off-by: Simon Hausmann <shausman@trolltech.com>
Signed-off-by: Lars Knoll <lars@trolltech.com>